pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "meeting-scheduler-env"
version = "1.0.0"
description = "Meeting Scheduler RL Environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "openai>=1.0.0",
    "openenv-core>=0.2.0",
    "requests>=2.31.0",
]

[project.scripts]
server = "server.app:main"
